api: Change int to float.

+++ VisualPinball.Engine.PinMAME.Unity/Runtime/PinMameGamelogicEngine.cs
@@ -620,7 +620,7 @@ namespace VisualPinball.Engine.PinMAME
 			OnCoilChanged?.Invoke(this, new CoilEventArgs(n, value));
 		}
 
-		public void SetLamp(string id, int value, bool isCoil = false, LampSource source = LampSource.Lamp)
+		public void SetLamp(string id, float value, bool isCoil = false, LampSource source = LampSource.Lamp)
 		{
 			OnLampChanged?.Invoke(this, new LampEventArgs(id, value, isCoil, source));
 		}
